Hbot/requirements.txt
pandas
langchain
langchain-community
chromadb
fastapi
uvicorn
python-telegram-bot[ext]